Exercise 2 Group Work Contribution
I copied the processingfile.qmd incorrectly as an .rmd file first, so I copied it again into a .qmd file before updating the code and cleaning the new data. I also copied the eda.qmd file incorrectly as an .rmd file originally, but there is now a correct eda2.qmd with the code for the 2 new figures. Both figures are saved in the figures folder found inside the results folder.